+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Filtering by language on queues list.
+- Service information on queues list.
+
 ## [v20.3.1](https://github.com/lexicalunit/spellbot/releases/tag/v20.3.1) - 2026-05-26
 
 ### Added